requirements.txt
torch>=2.0.0
transformers>=5.0.0
accelerate
gradio>=4.0.0
Pillow
numpy
faiss-cpu
